+++ client/src/pages/Account/graphql/graphql.ts
@@ -59,21 +59,3 @@ export const GET_ACCOUNTS = gql`
     }
   }
 `;
-
-export const GET_POSITIONS = gql`
-  query GetPositions($userId: String!, $aggregateId: String!) {
-    getPositions(userId: $userId, aggregateId: $aggregateId) {
-      userId
-      sk
-      aggregateId
-      symbol
-      description
-      exchange
-      currency
-      shares
-      acb
-      bookValue
-      marketValue
-    }
-  }
-`;
